[Refactor] #12 - separate SlideCard from CardView
